requirements.txt

python-dotenv
pandas
rapidfuzz
matplotlib
datetime